feat: adding feature to save transaction and build some exceptions errors features

+++ src/main/java/br/transaction/control/core/model/OperationType.java
@@ -7,11 +7,11 @@ import lombok.Getter;
 @Getter
 public enum OperationType {
 
-    COMPRA_A_VISTA(1),
-    COMPRA_PARCELADA(2),
-    SAQUE(3),
-    PAGAMENTO(4);
+    COMPRA_A_VISTA(1L),
+    COMPRA_PARCELADA(2L),
+    SAQUE(3L),
+    PAGAMENTO(4L);
 
-    private final Integer type;
+    private final Long type;
 
 }
